Changed all angle calculations to double instead of float, improved ESP, added more options for aimbot

+++ src/main/kotlin/com/charlatano/utils/Angle.kt
@@ -23,30 +23,30 @@ import java.lang.Math.abs
 typealias Angle = Vector
 
 fun Vector.normalize() = apply {
-	if (x != x) x = 0F
-	if (y != y) y = 0F
+	if (x != x) x = 0.0
+	if (y != y) y = 0.0
 
-	if (x > 89) x = 89F
-	if (x < -89) x = -89F
+	if (x > 89) x = 89.0
+	if (x < -89) x = -89.0
 
 	while (y > 180) y -= 360
 	while (y <= -180) y += 360
 
-	if (y > 180) y = 180F
-	if (y < -180F) y = -180F
+	if (y > 180) y = 180.0
+	if (y < -180F) y = -180.0
 
-	z = 0F
+	z = 0.0
 }
 
 internal fun Angle.distanceTo(target: Angle) = abs(x - target.x) + abs(y - target.y) + abs(z - target.z)
 
-internal fun Angle.finalize(orig: Angle, smoothing: Float) {
+internal fun Angle.finalize(orig: Angle, smoothing: Double) {
 	x -= orig.x
 	y -= orig.y
-	z = 0F
+	z = 0.0
 	normalize()
 
-	x = orig.x + x / 100F * smoothing
-	y = orig.y + y / 100F * smoothing
+	x = orig.x + x / 100.0 * smoothing
+	y = orig.y + y / 100.0 * smoothing
 	normalize()
 }
